made stories work for planner taskview and taskscell, added description to appointment cells
