--- web/src/components/TasksCell/TasksCell.mock.js
@@ -1,93 +1,0 @@
-// Define your own mock data here:
-export const standard = () => ({
-  tasks: [
-    {
-      id: 42,
-      title: 'Mock Task 1',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 3,
-      completed: 'not started',
-      duedate: new Date('2022-11-14T03:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 43,
-      title: 'Mock Task 2',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 1,
-      completed: 'not started',
-      duedate: new Date('2022-11-14T18:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 44,
-      title: 'Mock Task 3',
-      description: 'Mock Task 3 Description',
-      duration: null,
-      priority: 2,
-      completed: 'not started',
-      duedate: new Date('2022-11-13T03:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 45,
-      title: 'Mock Task 4',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 1,
-      completed: 'not started',
-      duedate: new Date('2022-11-12T03:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 46,
-      title: 'Mock Task 1',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 3,
-      completed: 'not started',
-      duedate: new Date('2022-11-14T03:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 47,
-      title: 'Mock Task 1',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 2,
-      completed: 'not started',
-      duedate: new Date('2022-11-12T03:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 48,
-      title: 'Mock Task 1',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 1,
-      completed: 'not started',
-      duedate: new Date('2022-11-11T15:24:00'),
-      createdat: null,
-      appointment: false
-    },
-    {
-      id: 49,
-      title: 'Mock Task 1',
-      description: 'Mock Task 2 Description',
-      duration: null,
-      priority: 3,
-      completed: 'not started',
-      duedate: new Date('2022-11-14T00:24:00'),
-      createdat: null,
-      appointment: false
-    },
-  ],
-})
